+++ packages/clerk-js/src/ui/components/SignIn/SignInStart.tsx
@@ -1,6 +1,6 @@
 import { useClerk } from '@clerk/shared/react';
 import { isWebAuthnAutofillSupported, isWebAuthnSupported } from '@clerk/shared/webauthn';
-import type { ClerkAPIError, SignInCreateParams } from '@clerk/types';
+import type { ClerkAPIError, SignInCreateParams, SignInResource } from '@clerk/types';
 import { useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react';
 
 import { ERROR_CODES } from '../../../core/constants';
@@ -231,18 +231,43 @@ export function _SignInStart(): JSX.Element {
 
   const buildSignInParams = (fields: Array<FormControlState<string>>): SignInCreateParams => {
     const hasPassword = fields.some(f => f.name === 'password' && !!f.value);
-    if (!hasPassword) {
+
+    /**
+     * FAPI will return an error when password is submitted but the user's email matches requires SAML authentication.
+     * We need to strip password from the create request, and reconstruct it later.
+     */
+    if (!hasPassword || userSettings.saml.enabled) {
       fields = fields.filter(f => f.name !== 'password');
     }
     return {
       ...buildRequest(fields),
-      ...(hasPassword && { strategy: 'password' }),
+      ...(hasPassword && !userSettings.saml.enabled && { strategy: 'password' }),
     } as SignInCreateParams;
+  };
+
+  const safePasswordSignInForSamlInstance = (
+    signInCreatePromise: Promise<SignInResource>,
+    fields: Array<FormControlState<string>>,
+  ) => {
+    return signInCreatePromise.then(signInResource => {
+      if (!userSettings.saml.enabled) {
+        return signInResource;
+      }
+      /**
+       * For SAML enabled instances, perform sign in with password only when it is allowed for the identified user.
+       */
+      const passwordField = fields.find(f => f.name === 'password')?.value;
+      if (!passwordField || signInResource.supportedFirstFactors.some(ff => ff.strategy === 'saml')) {
+        return signInResource;
+      }
+      return signInResource.attemptFirstFactor({ strategy: 'password', password: passwordField });
+    });
   };
 
   const signInWithFields = async (...fields: Array<FormControlState<string>>) => {
     try {
-      const res = await signIn.create(buildSignInParams(fields));
+      const res = await safePasswordSignInForSamlInstance(signIn.create(buildSignInParams(fields)), fields);
+
       switch (res.status) {
         case 'needs_identifier':
           // Check if we need to initiate a saml flow