+++ packages/ui/src/common/email-field.tsx
@@ -4,10 +4,14 @@ import React from 'react';
 import * as Field from '../primitives/field';
 
 export function EmailField({
-  className,
+  alternativeFieldTrigger,
   label = 'Email address',
+  hintText = 'Optional',
   ...props
-}: { label?: React.ReactNode } & Omit<React.ComponentProps<typeof Common.Input>, 'type'>) {
+}: { alternativeFieldTrigger?: React.ReactNode; label?: React.ReactNode; hintText?: string } & Omit<
+  React.ComponentProps<typeof Common.Input>,
+  'type'
+>) {
   return (
     <Common.Field
       name='emailAddress'
@@ -15,14 +19,20 @@ export function EmailField({
     >
       <Field.Root>
         <Common.Label asChild>
-          <Field.Label>{label}</Field.Label>
+          <Field.Label>
+            {label}{' '}
+            {alternativeFieldTrigger ? (
+              <Field.LabelEnd>{alternativeFieldTrigger}</Field.LabelEnd>
+            ) : !props?.required ? (
+              <Field.Hint>{hintText}</Field.Hint>
+            ) : null}
+          </Field.Label>
         </Common.Label>
         <Common.FieldState>
           {({ state }) => {
             return (
               <Common.Input
                 type='email'
-                className={className}
                 {...props}
                 asChild
               >