+++ packages/ui/src/common/phone-number-field/index.tsx
@@ -1,6 +1,6 @@
 import * as Common from '@clerk/elements/common';
-import cn from 'clsx';
 import { Command } from 'cmdk';
+import { cx } from 'cva';
 import * as React from 'react';
 import { Button, Dialog, DialogTrigger, Popover } from 'react-aria-components';
 
@@ -19,10 +19,16 @@ const countryOptions = Array.from(IsoToCountryMap.values()).map(country => {
 
 export const PhoneNumberField = React.forwardRef(function PhoneNumberField(
   {
+    alternativeFieldTrigger,
+    label = 'Phone number',
+    hintText = 'Optional',
     locationBasedCountryIso,
     onChange,
     ...props
   }: React.InputHTMLAttributes<HTMLInputElement> & {
+    alternativeFieldTrigger?: React.ReactNode;
+    label?: React.ReactNode;
+    hintText?: string;
     locationBasedCountryIso: CountryIso;
   },
   forwardedRef: React.ForwardedRef<HTMLInputElement>,
@@ -84,14 +90,21 @@ export const PhoneNumberField = React.forwardRef(function PhoneNumberField(
     >
       <Field.Root>
         <Common.Label asChild>
-          <Field.Label>Phone number</Field.Label>
+          <Field.Label>
+            {label}{' '}
+            {alternativeFieldTrigger ? (
+              <Field.LabelEnd>{alternativeFieldTrigger}</Field.LabelEnd>
+            ) : !props?.required ? (
+              <Field.Hint>{hintText}</Field.Hint>
+            ) : null}
+          </Field.Label>
         </Common.Label>
         <Common.FieldState>
           {({ state: intent }) => {
             return (
               <div
                 ref={containerRef}
-                className={cn(
+                className={cx(
                   'text-gray-12 border-gray-a6 flex w-full rounded-md border bg-white bg-clip-padding text-base outline-none',
                   'focus-within:ring-[0.1875rem] has-[[data-field-input][disabled]]:cursor-not-allowed has-[[data-field-input][disabled]]:opacity-50',
                   // intent