+++ packages/ui/src/components/sign-up/sign-up.tsx
@@ -1,7 +1,8 @@
 import { useClerk } from '@clerk/clerk-react';
 import * as Common from '@clerk/elements/common';
 import * as SignUp from '@clerk/elements/sign-up';
-import type { EnvironmentResource } from '@clerk/types';
+import { enUS } from '@clerk/localizations';
+import type { ClerkOptions, EnvironmentResource } from '@clerk/types';
 
 import { EmailField } from '~/common/email-field';
 import { FirstNameField } from '~/common/first-name-field';
@@ -11,13 +12,15 @@ import { PasswordField } from '~/common/password-field';
 import { PhoneNumberField } from '~/common/phone-number-field';
 import { UsernameField } from '~/common/username-field';
 import { PROVIDERS } from '~/constants';
+import { Alert } from '~/primitives/alert';
 import { Button } from '~/primitives/button';
 import * as Card from '~/primitives/card';
 import * as Connection from '~/primitives/connection';
 import * as Icon from '~/primitives/icon';
+import { LinkButton } from '~/primitives/link-button';
 import { Seperator } from '~/primitives/seperator';
-import { TextButton } from '~/primitives/text-button';
 import { getEnabledSocialConnectionsFromEnvironment } from '~/utils/getEnabledSocialConnectionsFromEnvironment';
+import { makeLocalizeable } from '~/utils/makeLocalizable';
 
 export function SignUpComponent() {
   return (
@@ -29,17 +32,21 @@ export function SignUpComponent() {
 
 function SignUpComponentLoaded() {
   const clerk = useClerk();
+  // TODO: Replace `any` with proper types
+  const t = makeLocalizeable(((clerk as any)?.options as ClerkOptions)?.localization || enUS);
   const enabledConnections = getEnabledSocialConnectionsFromEnvironment(
     (clerk as any)?.__unstable__environment as EnvironmentResource,
   );
   const locationBasedCountryIso = (clerk as any)?.__internal_country;
   const attributes = ((clerk as any)?.__unstable__environment as EnvironmentResource)?.userSettings.attributes;
+  const displayConfig = ((clerk as any)?.__unstable__environment as EnvironmentResource)?.displayConfig;
   const { enabled: firstNameEnabled, required: firstNameRequired } = attributes['first_name'];
   const { enabled: lastNameEnabled, required: lastNameRequired } = attributes['last_name'];
   const { enabled: usernameEnabled, required: usernameRequired } = attributes['username'];
   const { enabled: phoneNumberEnabled, required: phoneNumberRequired } = attributes['phone_number'];
   const { enabled: emailAddressEnabled, required: emailAddressRequired } = attributes['email_address'];
   const { enabled: passwordEnabled, required: passwordRequired } = attributes['password'];
+  const { applicationName, homeUrl, logoImageUrl } = displayConfig;
   return (
     <Common.Loading>
       {isGlobalLoading => {
@@ -49,10 +56,26 @@ function SignUpComponentLoaded() {
               <Card.Root>
                 <Card.Content>
                   <Card.Header>
-                    <Card.Title>Create your account</Card.Title>
-                    <Card.Description>Welcome! Please fill in the details to get started.</Card.Description>
+                    {logoImageUrl ? (
+                      <Card.Logo
+                        href={homeUrl}
+                        src={logoImageUrl}
+                        alt={applicationName}
+                      />
+                    ) : null}
+                    <Card.Title>{t('signUp.start.title')}</Card.Title>
+                    <Card.Description>
+                      {t('signUp.start.subtitle', {
+                        applicationName,
+                      })}
+                    </Card.Description>
                   </Card.Header>
                   <Card.Body>
+                    <Common.GlobalError>
+                      {({ message }) => {
+                        return <Alert>{message}</Alert>;
+                      }}
+                    </Common.GlobalError>
                     <Connection.Root>
                       {enabledConnections.map(c => {
                         const connection = PROVIDERS.find(provider => provider.id === c.provider);
@@ -85,16 +108,20 @@ function SignUpComponentLoaded() {
                       })}
                     </Connection.Root>
 
-                    <Seperator>or</Seperator>
+                    <Seperator>{t('dividerText')}</Seperator>
 
                     <div className='flex flex-col gap-4'>
                       {firstNameEnabled && lastNameEnabled ? (
                         <div className='flex gap-4'>
                           <FirstNameField
+                            label={t('formFieldLabel__firstName')}
+                            hintText={t('formFieldHintText__optional')}
                             required={firstNameRequired}
                             disabled={isGlobalLoading}
                           />
                           <LastNameField
+                            label={t('formFieldLabel__lastName')}
+                            hintText={t('formFieldHintText__optional')}
                             required={lastNameRequired}
                             disabled={isGlobalLoading}
                           />
@@ -103,28 +130,35 @@ function SignUpComponentLoaded() {
 
                       {usernameEnabled ? (
                         <UsernameField
+                          label={t('formFieldLabel__username')}
+                          hintText={t('formFieldHintText__optional')}
                           required={usernameRequired}
+                          disabled={isGlobalLoading}
+                        />
+                      ) : null}
+
+                      {emailAddressEnabled ? (
+                        <EmailField
+                          label={t('formFieldLabel__emailAddress')}
+                          hintText={t('formFieldHintText__optional')}
+                          required={emailAddressRequired}
                           disabled={isGlobalLoading}
                         />
                       ) : null}
 
                       {phoneNumberEnabled ? (
                         <PhoneNumberField
+                          label={t('formFieldLabel__phoneNumber')}
+                          hintText={t('formFieldHintText__optional')}
                           required={phoneNumberRequired}
                           disabled={isGlobalLoading}
                           locationBasedCountryIso={locationBasedCountryIso}
                         />
                       ) : null}
 
-                      {emailAddressEnabled ? (
-                        <EmailField
-                          required={emailAddressRequired}
-                          disabled={isGlobalLoading}
-                        />
-                      ) : null}
-
                       {passwordEnabled && passwordRequired ? (
                         <PasswordField
+                          label={t('formFieldLabel__password')}
                           required={passwordRequired}
                           disabled={isGlobalLoading}
                         />
@@ -143,7 +177,7 @@ function SignUpComponentLoaded() {
                               busy={isSubmitting}
                               disabled={isGlobalLoading || isSubmitting}
                             >
-                              Continue
+                              {t('formButtonPrimary')}
                             </Button>
                           </SignUp.Action>
                         );
@@ -154,7 +188,8 @@ function SignUpComponentLoaded() {
                 <Card.Footer>
                   <Card.FooterAction>
                     <Card.FooterActionText>
-                      Have an account? <Card.FooterActionLink href='/sign-in'>Sign in</Card.FooterActionLink>
+                      {t('signUp.start.actionText')}{' '}
+                      <Card.FooterActionLink href='/sign-in'>{t('signUp.start.actionLink')}</Card.FooterActionLink>
                     </Card.FooterActionText>
                   </Card.FooterAction>
                 </Card.Footer>
@@ -166,25 +201,54 @@ function SignUpComponentLoaded() {
                 <Card.Content>
                   <SignUp.Strategy name='phone_code'>
                     <Card.Header>
-                      <Card.Title>Verify your phone</Card.Title>
-                      <Card.Description>Enter the verification code sent to your phone</Card.Description>
+                      <Card.Title>{t('signUp.phoneCode.title')}</Card.Title>
+                      <Card.Description>{t('signUp.phoneCode.subtitle')}</Card.Description>
+                      <Card.Description>
+                        <span className='flex items-center justify-center gap-2'>
+                          {/* TODO: elements work
+                                    1. https://linear.app/clerk/issue/SDK-1830/add-signup-elements-for-accessing-email-address-and-phone-number
+                                    2. https://linear.app/clerk/issue/SDK-1831/pre-populate-emailphone-number-fields-when-navigating-back-to-the
+                          */}
+                          +1 (424) 424-4242{' '}
+                          <SignUp.Action
+                            navigate='start'
+                            asChild
+                          >
+                            <button
+                              type='button'
+                              className='focus-visible:ring-default size-4 rounded-sm outline-none focus-visible:ring-2'
+                              aria-label='Edit phone number'
+                            >
+                              <Icon.PencilUnderlined />
+                            </button>
+                          </SignUp.Action>
+                        </span>
+                      </Card.Description>
                     </Card.Header>
                     <Card.Body>
+                      <Common.GlobalError>
+                        {({ message }) => {
+                          return <Alert>{message}</Alert>;
+                        }}
+                      </Common.GlobalError>
                       <OTPField
                         disabled={isGlobalLoading}
                         // TODO:
-                        // 1. Replace `button` with `SignIn.Action` when `exampleMode` is removed
-                        // 2. Replace `button` with consolidated styles (tackled later)
+                        // 1. Replace `button` with consolidated styles (tackled later)
                         resend={
-                          <>
-                            Didn&apos;t recieve a code?{' '}
-                            <button
-                              type='button'
-                              className='text-accent-9 focus-visible:ring-default -mx-0.5 rounded-sm px-0.5 font-medium outline-none hover:underline focus-visible:ring-2'
-                            >
-                              Resend
-                            </button>
-                          </>
+                          <SignUp.Action
+                            asChild
+                            resend
+                            // eslint-disable-next-line react/no-unstable-nested-components
+                            fallback={({ resendableAfter }) => (
+                              <p className='text-gray-11 border border-transparent px-2.5 py-1.5 text-center text-base font-medium'>
+                                {t('signUp.phoneCode.resendButton')} (
+                                <span className='tabular-nums'>{resendableAfter}</span>)
+                              </p>
+                            )}
+                          >
+                            <LinkButton type='button'>{t('signUp.phoneCode.resendButton')}</LinkButton>
+                          </SignUp.Action>
                         }
                       />
                       <Common.Loading scope='step:verifications'>
@@ -199,7 +263,7 @@ function SignUpComponentLoaded() {
                                 disabled={isGlobalLoading}
                                 icon={<Icon.CaretRight />}
                               >
-                                Continue
+                                {t('formButtonPrimary')}
                               </Button>
                             </SignUp.Action>
                           );
@@ -210,25 +274,54 @@ function SignUpComponentLoaded() {
 
                   <SignUp.Strategy name='email_code'>
                     <Card.Header>
-                      <Card.Title>Verify your email</Card.Title>
-                      <Card.Description>Enter the verification code sent to your email</Card.Description>
+                      <Card.Title>{t('signUp.emailCode.title')}</Card.Title>
+                      <Card.Description>{t('signUp.emailCode.subtitle')}</Card.Description>
+                      <Card.Description>
+                        <span className='flex items-center justify-center gap-2'>
+                          {/* TODO: elements work
+                                    1. https://linear.app/clerk/issue/SDK-1830/add-signup-elements-for-accessing-email-address-and-phone-number
+                                    2. https://linear.app/clerk/issue/SDK-1831/pre-populate-emailphone-number-fields-when-navigating-back-to-the
+                          */}
+                          alex.carpenter@clerk.dev{' '}
+                          <SignUp.Action
+                            navigate='start'
+                            asChild
+                          >
+                            <button
+                              type='button'
+                              className='focus-visible:ring-default size-4 rounded-sm outline-none focus-visible:ring-2'
+                              aria-label='Edit email address'
+                            >
+                              <Icon.PencilUnderlined />
+                            </button>
+                          </SignUp.Action>
+                        </span>
+                      </Card.Description>
                     </Card.Header>
                     <Card.Body>
+                      <Common.GlobalError>
+                        {({ message }) => {
+                          return <Alert>{message}</Alert>;
+                        }}
+                      </Common.GlobalError>
                       <OTPField
                         disabled={isGlobalLoading}
                         // TODO:
-                        // 1. Replace `button` with `SignIn.Action` when `exampleMode` is removed
-                        // 2. Replace `button` with consolidated styles (tackled later)
+                        // 1. Replace `button` with consolidated styles (tackled later)
                         resend={
-                          <>
-                            Didn&apos;t recieve a code?{' '}
-                            <button
-                              type='button'
-                              className='text-accent-9 focus-visible:ring-default -mx-0.5 rounded-sm px-0.5 font-medium outline-none hover:underline focus-visible:ring-2'
-                            >
-                              Resend
-                            </button>
-                          </>
+                          <SignUp.Action
+                            asChild
+                            resend
+                            // eslint-disable-next-line react/no-unstable-nested-components
+                            fallback={({ resendableAfter }) => (
+                              <p className='text-gray-11 border border-transparent px-2.5 py-1.5 text-center text-base font-medium'>
+                                {t('signUp.emailCode.resendButton')} (
+                                <span className='tabular-nums'>{resendableAfter}</span>)
+                              </p>
+                            )}
+                          >
+                            <LinkButton type='button'>{t('signUp.emailCode.resendButton')}</LinkButton>
+                          </SignUp.Action>
                         }
                       />
                       <Common.Loading scope='step:verifications'>
@@ -243,7 +336,7 @@ function SignUpComponentLoaded() {
                                 disabled={isGlobalLoading}
                                 icon={<Icon.CaretRight />}
                               >
-                                Continue
+                                {t('formButtonPrimary')}
                               </Button>
                             </SignUp.Action>
                           );
@@ -254,24 +347,33 @@ function SignUpComponentLoaded() {
 
                   <SignUp.Strategy name='email_link'>
                     <Card.Header>
-                      <Card.Title>Verify your email</Card.Title>
-                      <Card.Description>Use the verification link sent to your email address</Card.Description>
+                      <Card.Title>{t('signUp.emailLink.title')}</Card.Title>
+                      <Card.Description>
+                        {t('signUp.emailLink.subtitle', {
+                          applicationName,
+                        })}
+                      </Card.Description>
                     </Card.Header>
                     <Card.Body>
+                      <Common.GlobalError>
+                        {({ message }) => {
+                          return <Alert>{message}</Alert>;
+                        }}
+                      </Common.GlobalError>
                       <SignUp.Action
                         resend
                         asChild
                         // eslint-disable-next-line react/no-unstable-nested-components
                         fallback={({ resendableAfter }) => {
                           return (
-                            <p className='text-gray-a11 focus-visible:ring-gray-a3 focus-visible:border-gray-a8 flex w-full items-center justify-center rounded-md border border-transparent bg-transparent bg-clip-padding px-2.5 py-1.5 text-base font-medium outline-none focus-visible:ring-[0.1875rem]'>
-                              Didn&apos;t recieve a link? Resend (
+                            <p className='text-gray-11 border border-transparent px-2.5 py-1.5 text-center text-base font-medium'>
+                              {t('signUp.emailLink.resendButton')} (
                               <span className='tabular-nums'>{resendableAfter}</span>)
                             </p>
                           );
                         }}
                       >
-                        <TextButton>Didn&apos;t recieve a link? Resend</TextButton>
+                        <LinkButton type='button'>{t('signUp.emailLink.resendButton')}</LinkButton>
                       </SignUp.Action>
                     </Card.Body>
                   </SignUp.Strategy>
@@ -284,18 +386,27 @@ function SignUpComponentLoaded() {
               <Card.Root>
                 <Card.Content>
                   <Card.Header>
-                    <Card.Title>Fill in missing fields</Card.Title>
-                    <Card.Description>Please fill in the remaining details to continue.</Card.Description>
+                    <Card.Title>{t('signUp.continue.title')}</Card.Title>
+                    <Card.Description>{t('signUp.continue.subtitle')}</Card.Description>
                   </Card.Header>
                   <Card.Body>
+                    <Common.GlobalError>
+                      {({ message }) => {
+                        return <Alert>{message}</Alert>;
+                      }}
+                    </Common.GlobalError>
                     <div className='space-y-4'>
                       {firstNameEnabled && lastNameEnabled ? (
                         <div className='flex gap-4'>
                           <FirstNameField
+                            label={t('formFieldLabel__firstName')}
+                            hintText={t('formFieldHintText__optional')}
                             required={firstNameRequired}
                             disabled={isGlobalLoading}
                           />
                           <LastNameField
+                            label={t('formFieldLabel__lastName')}
+                            hintText={t('formFieldHintText__optional')}
                             required={lastNameRequired}
                             disabled={isGlobalLoading}
                           />
@@ -304,6 +415,8 @@ function SignUpComponentLoaded() {
 
                       {usernameEnabled ? (
                         <UsernameField
+                          label={t('formFieldLabel__username')}
+                          hintText={t('formFieldHintText__optional')}
                           required={usernameRequired}
                           disabled={isGlobalLoading}
                         />
@@ -311,6 +424,8 @@ function SignUpComponentLoaded() {
 
                       {phoneNumberEnabled ? (
                         <PhoneNumberField
+                          label={t('formFieldLabel__phoneNumber')}
+                          hintText={t('formFieldHintText__optional')}
                           required={phoneNumberRequired}
                           disabled={isGlobalLoading}
                           locationBasedCountryIso={locationBasedCountryIso}
@@ -319,6 +434,8 @@ function SignUpComponentLoaded() {
 
                       {emailAddressEnabled ? (
                         <EmailField
+                          label={t('formFieldLabel__emailAddress')}
+                          hintText={t('formFieldHintText__optional')}
                           required={emailAddressRequired}
                           disabled={isGlobalLoading}
                         />
@@ -326,6 +443,7 @@ function SignUpComponentLoaded() {
 
                       {passwordEnabled && passwordRequired ? (
                         <PasswordField
+                          label={t('formFieldLabel__password')}
                           required={passwordRequired}
                           disabled={isGlobalLoading}
                         />
@@ -344,7 +462,7 @@ function SignUpComponentLoaded() {
                               busy={isSubmitting}
                               disabled={isGlobalLoading || isSubmitting}
                             >
-                              Continue
+                              {t('formButtonPrimary')}
                             </Button>
                           </SignUp.Action>
                         );
@@ -355,7 +473,8 @@ function SignUpComponentLoaded() {
                 <Card.Footer>
                   <Card.FooterAction>
                     <Card.FooterActionText>
-                      Have an account? <Card.FooterActionLink href='/sign-in'>Sign in</Card.FooterActionLink>
+                      {t('signUp.continue.actionText')}{' '}
+                      <Card.FooterActionLink href='/sign-in'>{t('signUp.continue.actionLink')}</Card.FooterActionLink>
                     </Card.FooterActionText>
                   </Card.FooterAction>
                 </Card.Footer>