+++ packages/upgrade/src/constants/versions.js
@@ -1,0 +1,4 @@
+export default [
+  { label: 'Core 2', value: 'core-2' },
+  { label: 'Core 1', value: 'core-1' },
+];